+++ CHANGELOG.md
@@ -41,7 +41,13 @@
 * Update Inventory: Set records for deletion permission's effect on ‘Set for deletion’ checkbox in Instance Edit view. Refs UIIN-3260.
 * *BREAKING* Migrate stripes dependencies to their Sunflower versions. Refs UIIN-3223.
 * *BREAKING* Migrate `react-intl` to v7. Refs UIIN-3224.
-* Item: Display all versions in View history second pane. Refs UIIN-3175.
+* Upgrade `browse` to `2.0`. Refs UIIN-3262.
+* *BREAKING* Use "number generator" for barcode, accession number and call number. Refs UIIN-2546.
+* *BREAKING* Instance: Display all versions in View history fourth pane. Refs UIIN-3173.
+* *BREAKING* MARC bib > View Source > Display Version History. Refs UIIN-3261.
+* "Copy barcode" icon is displayed next to item with no barcode. Fixes UIIN-3141.
+* *BREAKING* Holdings: Display all versions in View history second pane. Refs UIIN-3174.
+* *BREAKING* Item: Display all versions in View history second pane. Refs UIIN-3175.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)